Release 1.3.2
New: A button has been added to the JS Blocker UI for easy access to
switching between Expert and Simple mode.
New: Added some German strings.
Changed: When a button turns red when clicked on, it now pulsates to
help show that you must click again to initiate the action.
Fix: Resolves an issue wherein a hash change on a URL would immaturely
add inline frame data resulting in a false reading of the amount of
scripts blocked/allowed.
Fix: Clicking the divider in between Expanded and All in the rule
listing no longer deselects the left portion of the filter bar.
Fix: JS Blocker now works as expected on pages with special characters
in the URL.
Fix: Resolves issue with automatic rules not being disabled from the
main window when using the simple display.