Release 1.3.3
Changed: Pages that use custom created <iframe> elements (i.e. frames
created without the use of a "src" attribute) will now be displayed as
"(Non-URL Based Frame)" in the page drop down. This takes the place of
using the same URL as the main page as this was slightly misleading.
Rules modified with the new wording will still be based off of the URL
of the main page.
Fix: Using JS Blocker with multiple windows open now works as expected.
Fix: Closing a page in the background no longer messes up the display
of inline frames for the active page.
Fix: Improves compatibility with Safari 5.2 in relation to a possible
new way of storing window.location information.